+++ CHANGELOG.md
@@ -16,6 +16,7 @@ nav_order: 1
 - Add `make` targets `build` and `build-dev`
 - Add sweeper for account authentication
 - Make use of `BUILD_DEV_DIR` in `Makefile`
+- Capitalize variables in sample project
 
 ## [3.4.0] - 2022-07-26